update: grid,list render
