+++ js/clickGridCardList.js
@@ -4,6 +4,9 @@ import { renderGrid } from "./render/renderGrid.js";
 import { removeArrow } from "../utils/removeArrow.js";
 import Stores from "../utils/Store.js";
 
+const gridMain = document.getElementById("main-grid");
+const listMain = document.getElementById("main-list");
+
 const clickGridCardList = (logo, news) => {
   const clickGridImage = () => {
     const gridImage = document.getElementById("grid-image");
@@ -13,6 +16,8 @@ const clickGridCardList = (logo, news) => {
         "./img/card_list.svg"
       );
       changeImageSrc(e.target, "./img/clicked_grid.png");
+      gridMain.style.display = "grid";
+      listMain.style.display = "none";
       removeArrow();
       Stores.setPage(0);
       renderGrid(logo);
@@ -24,6 +29,8 @@ const clickGridCardList = (logo, news) => {
     cardListImage.addEventListener("click", (e) => {
       changeImageSrc(document.getElementById("grid-image"), "./img/grid.svg");
       changeImageSrc(e.target, "./img/clicked_card_list.png");
+      gridMain.style.display = "none";
+      listMain.style.display = "flex";
       removeArrow();
       Stores.setPage(0);
       renderCardList(news);
