+++ js/render/renderCardList.js
@@ -1,12 +1,9 @@
 import { drawInitCategory } from "../category.js";
 import { drawNews } from "../drawNews.js";
-import { gridMain, listMain } from "./renderMain.js";
 
-function renderCardList() {
-  gridMain.style.display = "none";
-  listMain.style.display = "flex";
+const renderCardList = (news) => {
   drawInitCategory();
   drawNews();
-}
+};
 
 export { renderCardList };
